# src/main/resources/swagger-def.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8443",
  "basePath": "/",
...
  "paths": {
...
    "/superhero/create": {
      "post": {
        "tags": [
          "Superhero Service REST API Interface"
        ],
        "summary": "Create a superhero and store it in application.",
        "description": "create a superhero and store it with all details in application.",
        "operationId": "createSuperheroUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*",
          "text/plain"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateSuperheroRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/superhero/get": {
      "get": {
        "tags": [
          "Superhero Service REST API Interface"
        ],
        "summary": "Pull a list of all superheroes stored in application.",
        "description": "Retrieve all superheroes details.",
        "operationId": "fetchAllSuperheroesUsingGET",
        "produces": [
          "*/*",
          "text/plain"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/superhero/get/{heroId}": {
      "get": {
        "tags": [
          "Superhero Service REST API Interface"
        ],
        "summary": "Find a superhero stored in application by superhero Id.",
        "description": "Find a superhero stored in application by superhero Id.",
        "operationId": "findByIdUsingGET",
        "produces": [
          "*/*",
          "text/plain"
        ],
        "parameters": [
          {
            "name": "heroId",
            "in": "path",
            "description": "heroId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/superhero/search/{heroName}": {
      "get": {
        "tags": [
          "Superhero Service REST API Interface"
        ],
        "summary": "Find a superhero stored in application by Name.",
        "description": "Find a superhero stored in application by superhero name.",
        "operationId": "findByNameUsingGET",
        "produces": [
          "*/*",
          "text/plain"
        ],
        "parameters": [
          {
            "name": "heroName",
            "in": "path",
            "description": "heroName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "CreateSuperheroRequest": {
      "type": "object",
      "required": [
        "allies",
        "skills"
      ],
      "properties": {
        "allies": {
          "type": "array",
          "description": "A collection of the returned superhero allies ID's",
          "allowEmptyValue": false,
          "items": {
            "type": "integer",
            "format": "int64"
          }
        },
        "firstAppearance": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "pseudonym": {
          "type": "string"
        },
        "publisher": {
          "type": "integer",
          "format": "int64"
        },
        "skills": {
          "type": "array",
          "description": "A collection of the returned superhero skills and powers ID's",
          "allowEmptyValue": false,
          "items": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "title": "CreateSuperheroRequest",
      "description": "Create Superhero Request resource representation"
    },
...
    "Publisher": {
      "type": "object",
      "properties": {
        "desc": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "Publisher"
    },
    "Response": {
      "type": "object",
      "required": [
        "allies",
        "skills",
        "superheroes"
      ],
      "properties": {
        "allies": {
          "type": "array",
          "description": "A collection of the returned superhero allies",
          "allowEmptyValue": false,
          "items": {
            "$ref": "#/definitions/Superhero"
          }
        },
        "skills": {
          "type": "array",
          "description": "A collection of the returned superhero skills and powers",
          "allowEmptyValue": false,
          "items": {
            "$ref": "#/definitions/Skill"
          }
        },
        "superhero": {
          "$ref": "#/definitions/Superhero"
        },
        "superheroes": {
          "type": "array",
          "description": "A collection of the returned superheroes details",
          "allowEmptyValue": false,
          "items": {
            "$ref": "#/definitions/Response"
          }
        }
      },
      "title": "Response",
      "description": "Response resource representation"
    },
    "Skill": {
      "type": "object",
      "properties": {
        "desc": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "Skill"
    },
    "Superhero": {
      "type": "object",
      "properties": {
        "firstAppearance": {
          "type": "string",
          "example": "yyyy-MM-dd"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "pseudonym": {
          "type": "string"
        },
        "publisher": {
          "$ref": "#/definitions/Publisher"
        }
      },
      "title": "Superhero"
    },
...
  }
}